feat: actions transactions + categories + banks and layout

+++ src/actions/categories/deleteCategories.actions.ts
@@ -1,3 +1,5 @@
+"use server"
+
 import { prisma } from "@/lib/prisma";
 import { getUserAuth } from "../users/getUserAuth.actions";
 
@@ -8,6 +10,21 @@ interface DeleteCategoriesProps {
 export async function deleteCategories({ categoryId }: DeleteCategoriesProps) {
   try {
     const userId = await getUserAuth();
+
+    const transactionsWithCategoryId = await prisma.transactions.count({
+      where: {
+        userId,
+        categoryId
+      }
+    })
+
+    if(transactionsWithCategoryId > 0) {
+      return {
+        error: new Error("Category has transactions"),
+        success: false,
+        message: "Category has transactions",
+      };
+    }
 
     await prisma.categories.delete({
       where: {
